refactor(product): change value type to BigDecimal

+++ src/main/java/com/olo/apimonolithic/feature/productmanagement/domain/model/Product.java
@@ -2,10 +2,12 @@ package com.olo.apimonolithic.feature.productmanagement.domain.model;
 
 import com.olo.apimonolithic.feature.categorymanagement.domain.model.Category;
 
+import java.math.BigDecimal;
+
 public record Product(
         Long id,
         String name,
         Category category,
-        Long value
+        BigDecimal value
 ) {
 }
